Create a joined network to visualize overlapping connections with the same outer joined node set.

Create a joined network to visualize overlapping connections with the same outer joined node set.

Usage

create_joined_networks(
  e,
  correlation_list_names = c("male_agg", "female_non"),
  channels = "cfos",
  ontology = "unified",
  alpha = 0.001,
  pearson_thresh = 0.9,
  proportional_thresh = NULL,
  alpha2 = NULL,
  pearson_thresh2 = NULL,
  proportional_thresh2 = NULL,
  anatomical.order = c("Isocortex", "OLF", "HPF", "CTXsp", "CNU", "TH", "HY", "MB", "HB",
    "CB"),
  export_overlapping_edges = TRUE
)

Arguments

e

experiment object

correlation_list_names

(str vec) character vector of the two correlation lists used to include in a joined network

channels

(str, default = c("cfos", "eyfp", "colabel")) The channels to process.

ontology

(str, default = "allen") Region ontology to use. options = "allen" or "unified"

alpha

(float, default = 0.05) The significance threshold for including brain regions in the network. if NULL or NA, this threshold is not applied.

pearson_thresh

(float, default = 0.8) The pearson correlation coefficient threshold to apply for filtering out

proportional_thresh

(float, default = NULL) Takes precedent over the alpha and the pearson_thresh parameters. Input the desired edge proportion (i.e., edge density) as your desired sparsity constraint.

alpha2

(NULL) If not NULL, this gives the option of filtering the second network by a different alpha from the first. The alpha parameter will then be used as the threshold for network 1.

pearson_thresh2

(NULL) If not NULL, this gives the option of filtering the second network by a different pearson threshold from the first network. The pearson_thresh parameter will then be used as the threshold for network 1.

proportional_thresh2

(NULL) If not NULL, this gives the option of filtering the second network by a different proportional threshold from the first.

anatomical.order

(vec, c("Isocortex","OLF","HPF","CTXsp","CNU","TH","HY","MB","HB","CB")) The default super region acronym list that groups all subregions in the dataset.

export_overlapping_edges

(bool, default = TRUE) Whether to export the overlapping edges between the two networks as a csv into the table directory.

Value

e experiment object. This object now has a new added element called networks. This is a list storing a graph object per channel for each network analysis run. The name of each network (network_name) is the same as the correlation_list_name

used to generate the network. This network_name is fed as a parameter into the plot_networks() function.

See also

Examples