Updated function to return character vector of the super region acronyms after inputting a character vector of acronyms

Updated function to return character vector of the super region acronyms after inputting a character vector of acronyms

Usage

get.super.regions(
  acronym,
  anatomical.order = c("Isocortex", "OLF", "HPF", "CTXsp", "CNU", "TH", "HY", "MB", "HB",
    "CB"),
  ontology = "allen"
)

Arguments

acronym

(str vec) character vector of acronyms. Cannot be a factor vector.

anatomical.order

(default = c("Isocortex", "OLF", "HPF", "CTXsp", "CNU","TH", "HY", "MB", "HB", "CB")) Default way to group subregions into super regions order

ontology

(str, default = "allen") Region ontology to use. options = "allen" or "unified"

Value

vector the same length as acronym consisting of the super regions of the acronyms

Examples

get.super.regions(acronym = c("DG-sg", "ACA5", "CLA", "ACB", "BLA", "PERI" , "PB"))
#> [1] "HPF"       "Isocortex" "CTXsp"     "CNU"       "CTXsp"     "Isocortex"
#> [7] "HB"       
get.super.regions(acronym = c("CPi", "MHb", "ZI", "Acb", "BL"), ontology = "unified")
#> [1] "CNU"   "TH"    "HY"    "CNU"   "CTXsp"