Perform proportional to population size (PPS) sampling based on SMART approach

Perform proportional to population size (PPS) sampling based on SMART approach

get_pps(df, pop, m)

Arguments

df
A data.frame of clusters/villages with their population size

pop
A character value specifying the name of the variable in `df` for the population size information

m
Number of clusters

Value

A data.frame which is a subset of `df` of the selected clusters for the survey sample. A new variable called `cluster_id` is created which identifies the cluster number and the reserved clusters.

Examples

get_pps(df = clusterList, pop = "pop", m = 30)

#>    cluster_id village_id              village_name  pop       cum_pop
#> 1          C1         12                     Ayata 1070     4306-5375
#> 2          C2         17                 Tinibilka  550   10596-11145
#> 3          C3         25      Berara tuwate tsehay  915   15446-16360
#> 4          C4         31        Fela Tuwate tsehay  960   21441-22400
#> 5          C5         43             Assinefechign  850   32116-32965
#> 6          C6         51                 Chewagisa  870   37356-38225
#> 7          C7         57                    Guassa 1270   41931-43200
#> 8          C8         67                     Derka 1020   47761-48780
#> 9          C9         76                    Gergas 1075   53530-54605
#> 10        C10         82                     Arero  610   58951-59560
#> 11        C11         95                   Betassa  751   69990-70740
#> 12        C12        102                      Jema  751   75496-76246
#> 13        C13        110               Dabosh Ager  515   80764-81278
#> 14        C14        126 Dolie Ager and Abaya ager 1030   86228-87258
#> 15        C15        133               Abageferssa  480   91524-92003
#> 16        C16        141                    Tigaja  805   97189-97993
#> 17        C17        154                    Jejero  514 102536-103050
#> 18        C18        162                     Yegof  885 107568-108452
#> 19        C19        168                   Bekimos  750 113113-113862
#> 20        C20        177                    Abecho  770 118623-119392
#> 21        C21        191          Doyenona Bukassa  720 129133-129852
#> 22        C22        200                  Abadeneb  450 134927-135377
#> 23        C23        210        Abiagerna Betoager  625 140488-141112
#> 24        C24        219              Lebona Derka  590 145553-146142
#> 25        C25        227                Dawud Ager  780 150753-151532
#> 26        C26        236                 Aba balie  735 156578-157312
#> 27        C27        245      Legelafetona Alebeko  645 162028-162672
#> 28        C28        265                    Gadula  630 172948-173577
#> 29        C29        274                     Washa  870 178048-178917
#> 30        C30        281                      Dego  945 183178-184122
#> 31        RC1        185                      Tiro  580 124203-124782
#> 32        RC2         89                 Werebecho  786   64460-65245
#> 33        RC3        255                Legelafeto  530 167278-167807
#> 34        RC4         37                    Meyane 1120   26901-28020