Get the cumulative population size range

Get the cumulative population size range

roll_sum_range(v, na = c("na.rm", "na.omit"), as_char = FALSE)

Arguments

v
A vector of population sizes to apply the cumulative population size range calculations on

na
How should missing values be treated? Either `na.rm` to remove or `na.omit` to omit by assigning a value of 0. Default is to omit.

as_char
Logical. Should output be a character vector or a data.frame of two values. Default to FALSE.

Value

A vector of cumulative population size range

Examples

roll_sum_range(v = clusterList$pop)

#>     lower_range upper_range
#> 1             1         530
#> 2           531         735
#> 3           736        1005
#> 4          1006        1265
#> 5          1266        1655
#> 6          1656        2000
#> 7          2001        2385
#> 8          2386        2710
#> 9          2711        3070
#> 10         3071        3270
#> 11         3271        4305
#> 12         4306        5375
#> 13         5376        6625
#> 14         6626        7775
#> 15         7776        9205
#> 16         9206       10595
#> 17        10596       11145
#> 18        11146       11980
#> 19        11981       12655
#> 20        12656       13490
#> 21        13491       13975
#> 22        13976       14390
#> 23        14391       14915
#> 24        14916       15445
#> 25        15446       16360
#> 26        16361       17530
#> 27        17531       18430
#> 28        18431       19380
#> 29        19381       20400
#> 30        20401       21440
#> 31        21441       22400
#> 32        22401       23525
#> 33        23526       24200
#> 34        24201       25325
#> 35        25326       25975
#> 36        25976       26900
#> 37        26901       28020
#> 38        28021       28855
#> 39        28856       30005
#> 40        30006       30605
#> 41        30606       31260
#> 42        31261       32115
#> 43        32116       32965
#> 44        32966       33615
#> 45        33616       34055
#> 46        34056       34495
#> 47        34496       35195
#> 48        35196       35800
#> 49        35801       36455
#> 50        36456       37355
#> 51        37356       38225
#> 52        38226       38920
#> 53        38921       39440
#> 54        39441       40180
#> 55        40181       41045
#> 56        41046       41930
#> 57        41931       43200
#> 58        43201       43540
#> 59        43541       44040
#> 60        44041       44345
#> 61        44346       44695
#> 62        44696       45790
#> 63        45791       46620
#> 64        46621       46970
#> 65        46971       47370
#> 66        47371       47760
#> 67        47761       48780
#> 68        48781       49110
#> 69        49111       49620
#> 70        49621       50135
#> 71        50136       50555
#> 72        50556       51635
#> 73        51636       52295
#> 74        52296       52985
#> 75        52986       53530
#> 76        53531       54605
#> 77        54606       55300
#> 78        55301       56280
#> 79        56281       57025
#> 80        57026       57815
#> 81        57816       58950
#> 82        58951       59560
#> 83        59561       60410
#> 84        60411       61315
#> 85        61316       62101
#> 86        62102       62887
#> 87        62888       63673
#> 88        63674       64459
#> 89        64460       65245
#> 90        65246       66031
#> 91        66032       66817
#> 92        66818       68417
#> 93        68418       69203
#> 94        69204       69989
#> 95        69990       70740
#> 96        70741       71740
#> 97        71741       72490
#> 98        72491       73242
#> 99        73243       73993
#> 100       73994       74744
#> 101       74745       75495
#> 102       75496       76246
#> 103       76247       76997
#> 104       76998       77748
#> 105       77749       78213
#> 106       78214       78978
#> 107       78979       79623
#> 108       79624       80288
#> 109       80289       80763
#> 110       80764       81278
#> 111       81279       81928
#> 112       81929       82803
#> 113       82804       83098
#> 114       83099       83253
#> 115       83254       83523
#> 116       83524       83713
#> 117       83714       83888
#> 118       83889       84038
#> 119       84039       84208
#> 120       84209       84363
#> 121       84364       84643
#> 122       84644       84983
#> 123       84984       85253
#> 124       85254       85643
#> 125       85644       86228
#> 126       86229       87258
#> 127       87259       87693
#> 128       87694       88198
#> 129       88199       88933
#> 130       88934       89743
#> 131       89744       90773
#> 132       90774       91523
#> 133       91524       92003
#> 134       92004       93183
#> 135       93184       93618
#> 136       93619       94238
#> 137       94239       94693
#> 138       94694       95413
#> 139       95414       96438
#> 140       96439       97188
#> 141       97189       97993
#> 142       97994       98698
#> 143       98699       99073
#> 144       99074       99323
#> 145       99324       99508
#> 146       99509       99953
#> 147       99954      100223
#> 148      100224      100618
#> 149      100619      101293
#> 150      101294      101788
#> 151      101789      102536
#> 152      102537      103050
#> 153      103051      103727
#> 154      103728      104429
#> 155      104430      105105
#> 156      105106      105877
#> 157      105878      106462
#> 158      106463      107125
#> 159      107126      107567
#> 160      107568      108452
#> 161      108453      109397
#> 162      109398      110367
#> 163      110368      111372
#> 164      111373      112242
#> 165      112243      113112
#> 166      113113      113862
#> 167      113863      114737
#> 168      114738      115247
#> 169      115248      115667
#> 170      115668      116237
#> 171      116238      116852
#> 172      116853      117387
#> 173      117388      117977
#> 174      117978      118622
#> 175      118623      119392
#> 176      119393      119977
#> 177      119978      120532
#> 178      120533      121422
#> 179      121423      122422
#> 180      122423      122942
#> 181      122943      123412
#> 182      123413      124202
#> 183      124203      124782
#> 184      124783      125707
#> 185      125708      126642
#> 186      126643      127662
#> 187      127663      128632
#> 188      128633      129132
#> 189      129133      129852
#> 190      129853      130692
#> 191      130693      131262
#> 192      131263      131762
#> 193      131763      132342
#> 194      132343      132752
#> 195      132753      133477
#> 196      133478      134347
#> 197      134348      134927
#> 198      134928      135377
#> 199      135378      135842
#> 200      135843      136357
#> 201      136358      137172
#> 202      137173      137672
#> 203      137673      138172
#> 204      138173      138737
#> 205      138738      139467
#> 206      139468      139892
#> 207      139893      140487
#> 208      140488      141112
#> 209      141113      141672
#> 210      141673      142437
#> 211      142438      143087
#> 212      143088      143682
#> 213      143683      144457
#> 214      144458      144992
#> 215      144993      145552
#> 216      145553      146142
#> 217      146143      146762
#> 218      146763      147402
#> 219      147403      148032
#> 220      148033      148562
#> 221      148563      149072
#> 222      149073      150032
#> 223      150033      150752
#> 224      150753      151532
#> 225      151533      152372
#> 226      152373      152977
#> 227      152978      153557
#> 228      153558      154137
#> 229      154138      154692
#> 230      154693      155397
#> 231      155398      155947
#> 232      155948      156577
#> 233      156578      157312
#> 234      157313      158032
#> 235      158033      158647
#> 236      158648      159227
#> 237      159228      159807
#> 238      159808      160242
#> 239      160243      160777
#> 240      160778      161407
#> 241      161408      162027
#> 242      162028      162672
#> 243      162673      163137
#> 244      163138      163572
#> 245      163573      164082
#> 246      164083      164607
#> 247      164608      165052
#> 248      165053      165742
#> 249      165743      166262
#> 250      166263      166742
#> 251      166743      167277
#> 252      167278      167807
#> 253      167808      168207
#> 254      168208      168777
#> 255      168778      169272
#> 256      169273      169822
#> 257      169823      170442
#> 258      170443      171087
#> 259      171088      171642
#> 260      171643      172187
#> 261      172188      172947
#> 262      172948      173577
#> 263      173578      174132
#> 264      174133      174622
#> 265      174623      175137
#> 266      175138      175672
#> 267      175673      176117
#> 268      176118      176562
#> 269      176563      177437
#> 270      177438      178047
#> 271      178048      178917
#> 272      178918      179812
#> 273      179813      180562
#> 274      180563      181147
#> 275      181148      181912
#> 276      181913      182702
#> 277      182703      183177
#> 278      183178      184122