Perform proportional to population size (PPS) sampling based on SMART approach

Perform proportional to population size (PPS) sampling based on SMART approach

get_pps(df, pop, m)

Arguments

df
A data.frame of clusters/villages with their population size

pop
A character value specifying the name of the variable in `df` for the population size information

m
Number of clusters

Value

A data.frame which is a subset of `df` of the selected clusters for the survey sample. A new variable called `cluster_id` is created which identifies the cluster number and the reserved clusters.

Examples

get_pps(df = villageList1, pop = "pop", m = 30)

#> $`Sample clusters 1`
#>     village_id              village_name  pop       cum_pop clusters
#> 1            1                      Libo  530         1-530         
#> 2            2                     Derka  205       531-735         
#> 3            3                     Abito  270      736-1005         
#> 4            4                    Yebiso  260     1006-1265         
#> 5            5                     Amebo  390     1266-1655         
#> 6            6                     Mille  345     1656-2000         
#> 7            7                   Abajala  385     2001-2385         
#> 8            8              Kutayuwedajo  325     2386-2710         
#> 9            9                    Abesho  360     2711-3070         
#> 10          10                 Shenbutie  200     3071-3270         
#> 11          11                      Jila 1035     3271-4305         
#> 12          12                     Ayata 1070     4306-5375        1
#> 13          13                  Werbecho 1250     5376-6625         
#> 14          14                   Yebello 1150     6626-7775         
#> 15          15              Mayan mamedo 1430     7776-9205         
#> 16          16             Mayan Ameraro 1390    9206-10595         
#> 17          17                 Tinibilka  550   10596-11145        2
#> 18          18                     Arado  835   11146-11980         
#> 19          19                 Alekodayo  675   11981-12655         
#> 20          20                  Weredero  835   12656-13490         
#> 21          21                    Metero  485   13491-13975         
#> 22          22                   Layamba  415   13976-14390         
#> 23          23                   Jerjero  525   14391-14915         
#> 24          24                    Wushel  530   14916-15445         
#> 25          25      Berara tuwate tsehay  915   15446-16360        3
#> 26          26        Berara mata tsehay 1170   16361-17530         
#> 27          27                   Tikissa  900   17531-18430         
#> 28          28                Adiss Amba  950   18431-19380         
#> 29          29                    Genfot 1020   19381-20400         
#> 30          30          Fela mata tsehay 1040   20401-21440         
#> 31          31        Fela Tuwate tsehay  960   21441-22400        4
#> 32          32                   Jerjero 1125   22401-23525         
#> 33          33                     Arado  675   23526-24200         
#> 34          34                      Koso 1125   24201-25325         
#> 35          35                Amorgedele  650   25326-25975         
#> 36          36                     Ketie  925   25976-26900         
#> 37          37                    Meyane 1120   26901-28020        5
#> 38          38                    Tedech  835   28021-28855         
#> 39          39                     Gobey 1150   28856-30005         
#> 40          40                      Jema  600   30006-30605         
#> 41          41                     Golbo  655   30606-31260         
#> 42          42               Laygodemesh  855   31261-32115         
#> 43          43             Assinefechign  850   32116-32965        6
#> 44          44                    Yebase  650   32965-33615         
#> 45          45                   Anegefa  440   33616-34055         
#> 46          46                    Chefie  440   34056-34495         
#> 47          47                 Buko Ager  700   34496-35195         
#> 48          48             Chorka mender  605   35196-35800         
#> 49          49              Abulie  Ager  655   35801-36455         
#> 50          50                    Megera  900   36456-37355         
#> 51          51                 Chewagisa  870   37356-38225        7
#> 52          52                  Kermemie  695   38226-38920         
#> 53          53                Layebelsie  520   38921-39440         
#> 54          54               Tachebelsie  740   39441-40180         
#> 55          55                Wedey Ager  865   40181-41045         
#> 56          56                     Eriko  885   41046-41930         
#> 57          57                    Guassa 1270   41931-43200        8
#> 58          58                 Maru Ager  340   43201-43540         
#> 59          59                   Temsase  500   43541-44040         
#> 60          60                   Werdero  305   44041-44345         
#> 61          61              Mehalejebene  350   44346-44695         
#> 62          62                     Weltu 1095   44695-45790         
#> 63          63              Mentawu Adem  830   45791-46620         
#> 64          64                    Mamedo  350   46621-46970         
#> 65          65                  Kombelti  400   46971-47370         
#> 66          66                Kedir Ager  390   47371-47760         
#> 67          67                     Derka 1020   47761-48780        9
#> 68          68                   Gelbite  330   48781-49110         
#> 69          69                  Neftegna  510   49111-49620         
#> 70          70               Mehalgedewa  515   49621-50135         
#> 71          71                   Gerjaja  420   50135-50555         
#> 72          72          Meskel Metekosha 1080   50556-51635         
#> 73          73                   Gobeden  660   51636-52295         
#> 74          74                 Alyu Ager  690   52296-52985         
#> 75          75                   Temsase  545   52986-53530         
#> 76          76                    Gergas 1075   53530-54605       10
#> 77          77                 Kili Ager  695   54606-55300         
#> 78          78         Ali Ager/Abu ager  980   55301-56280         
#> 79          79                    Gobeya  745   56281-57025         
#> 80          80               Abesho Ager  790   57026-57815         
#> 81          81                   Webebie 1135   57816-58950         
#> 82          82                     Arero  610   58951-59560       11
#> 83          83                 Lebo Ager  850   59561-60410         
#> 84          84                   Erkoche  905   60411-61315         
#> 85          85                     Wayna  786   61316-62101         
#> 86          86               Geshomender  786   62102-62887         
#> 87          87                 Dejuwasha  786   62888-63673         
#> 88          88             Abayasin jema  786   63673-64459         
#> 89          89                 Werebecho  786   64460-65245       12
#> 90          90                     Motie  786   65246-66031         
#> 91          91                  Keygedel  786   66032-66817         
#> 92          92                Tidegebeya 1600   66818-68417         
#> 93          93                   Mechela  786   68418-69203         
#> 94          94                  Chichita  786   69204-69989         
#> 95          95                   Betassa  751   69990-70740       13
#> 96          96                    Albiko 1000   70741-71740         
#> 97          97                   Jimetit  750   71741-72490         
#> 98          98                   Sefiyat  752   72491-73242         
#> 99          99                   Yejerti  751   73243-73993         
#> 100        100                    Chermu  751   73994-74744         
#> 101        101                     Alaye  751   74744-75495         
#> 102        102                      Jema  751   75496-76246       14
#> 103        103                   Derenso  751   76247-76997         
#> 104        104                    Jorkie  751   76998-77748         
#> 105        105                   Mukedad  465   77749-78213         
#> 106        106               AbabidaAger  765   78214-78978         
#> 107        107                    Yinale  645   78979-79623         
#> 108        108                  Ali Ager  665   79624-80288         
#> 109        109              Weldiye Ager  475   80289-80763         
#> 110        110               Dabosh Ager  515   80764-81278       15
#> 111        111                 Delo Ager  650   81279-81928         
#> 112        112                  Jimogodo  875   81929-82803         
#> 113        113        Ababilana karagora  295   82804-83098         
#> 114        114                  Wereruga  155   83099-83253         
#> 115        115          Abajeruna zobele  270   83254-83523         
#> 116        116                Chikaberet  190   83524-83713         
#> 117        117                 Abakkurie  175   83714-83888         
#> 118        118                 Abawussen  150   83889-84038         
#> 119        119                   Ababula  170   84039-84208         
#> 120        120                    Wujira  155   84209-84363         
#> 121        121                   Abajila  280   84364-84643         
#> 122        122                    Kelina  340   84643-84983         
#> 123        123   2 haraAbawatie,Aba keye  270   84984-85253         
#> 124        124              Abawdya Ager  390   85254-85643         
#> 125        125              Eshetie Ager  585   85644-86228         
#> 126        126 Dolie Ager and Abaya ager 1030   86228-87258       16
#> 127        127             Bejerond Ager  435   87258-87693         
#> 128        128               Mussie Ager  505   87694-88198         
#> 129        129               Tolaye Ager  735   88199-88933         
#> 130        130                Assen Ager  810   88934-89743         
#> 131        131               Kurach Ager 1030   89743-90773         
#> 132        132                       Kat  750   90774-91523         
#> 133        133               Abageferssa  480   91524-92003       17
#> 134        134                    Asgedo 1180   92004-93183         
#> 135        135                Abaweldiye  435   93183-93618         
#> 136        136                   Terfoch  620   93619-94238         
#> 137        137                   Gobeden  455   94239-94693         
#> 138        138                Adiss Amba  720   94694-95413         
#> 139        139        Sheh Agerna totela 1025   95414-96438         
#> 140        140              Kotuna Beret  750   96439-97188         
#> 141        141                    Tigaja  805   97189-97993       18
#> 142        142             Diblena senta  705   97994-98698         
#> 143        143                  Bonaager  375   98699-99073         
#> 144        144                    Acheno  250   99074-99323         
#> 145        145                 Abuyeager  185   99324-99508         
#> 146        146                      Leku  445   99509-99953         
#> 147        147                    Gedero  270  99954-100223         
#> 148        148                 Doyo Ager  395 100224-100618         
#> 149        149                   Mentera   NA        100618         
#> 150        150                     Samte   NA        100618         
#> 151        151                     Wedey  675 100619-101293         
#> 152        152                      Ellu  495 101294-101788         
#> 153        153                     Adela  748 101789-102536         
#> 154        154                    Jejero  514 102536-103050       19
#> 155        155                  Derenssa  677 103051-103727         
#> 156        156                    Matebo  702 103727-104429         
#> 157        157                   Kenbeto  676 104430-105105         
#> 158        158                   Korerie  772 105106-105877         
#> 159        159                    Negewo  585 105876-106462         
#> 160        160                    Haroye  663 106463-107125         
#> 161        161                      Hedo  442 107126-107567         
#> 162        162                     Yegof  885 107568-108452       20
#> 163        163                    Duleta  945 108452-109397         
#> 164        164                 Babu ager  970 109397-110367         
#> 165        165             Chorka Mender 1005 110367-111372         
#> 166        166                     Tedie  870 111373-112242         
#> 167        167                   Birbira  870 112243-113112         
#> 168        168                   Bekimos  750 113113-113862       21
#> 169        169                 Jewa Ager  875 113863-114737         
#> 170        170                  Abachufa  510 114738-115247         
#> 171        171                   Adengur  420 115248-115667         
#> 172        172                   Abalema  570 115668-116237         
#> 173        173              Amedu Mender  615 116238-116852         
#> 174        174                Timhirtbet  535 116853-117387         
#> 175        175                Shoto Ager  590 117389-117977         
#> 176        176                   Degendo  645 117978-118622         
#> 177        177                    Abecho  770 118623-119392       22
#> 178        178                   Salaysh  585 119393-119977         
#> 179        179                 Biru Ager  555 119978-120532         
#> 180        180                   Adarash  890 120533-121422         
#> 181        181                    Bilido 1000 121422-122422         
#> 182        182                    Kurkur  520 122423-122942         
#> 183        183                   Tebasit  470 122943-123412         
#> 184        184                      Jema  790 123413-124202         
#> 185        185                      Tiro  580 124203-124782       23
#> 186        186                Temujegola  925 124783-125707         
#> 187        187              Jegolasitaya  935 125708-126642         
#> 188        188           Kutayena Bayush 1020 126643-127662         
#> 189        189               Temuna Adey  970 127663-128632         
#> 190        190          Chanchona Weynie  500 128634-129132         
#> 191        191          Doyenona Bukassa  720 129133-129852       24
#> 192        192                    Gobeya  840 129853-130692         
#> 193        193          Kutana Tide Amba  570 130693-131262         
#> 194        194                    Gelsha  500 131263-131762         
#> 195        195               Tach Gelsha  580 131763-132342         
#> 196        196                 Asho Ager  410 132343-132752         
#> 197        197        Tofina Giragn meda  725 132753-133477         
#> 198        198            Mendebona feto  870 133478-134347         
#> 199        199                    Kutaye  580 134348-134927         
#> 200        200                  Abadeneb  450 134927-135377       25
#> 201        201              Yererie gora  465 135378-135842         
#> 202        202                     Kolbo  515 135843-136357         
#> 203        203                    Wayana  815 136358-137172         
#> 204        204               Gena Memcha  500 137173-137672         
#> 205        205           Kundina Yekosso  500 137672-138172         
#> 206        206            Aba Jigar Ager  565 138173-138737         
#> 207        207                    Wateya  730 138738-139467         
#> 208        208          Negwelina Gotera  425 139468-139892         
#> 209        209           Ginbarona gudfo  595 139893-140487         
#> 210        210        Abiagerna Betoager  625 140488-141112       26
#> 211        211            Jema ababorena   NA        141112         
#> 212        212          Nechirona Jareso  560 141113-141672         
#> 213        213             Sukina sekela  765 141673-142437         
#> 214        214        Ashoyena Arbachama  650 142438-143087         
#> 215        215               Habona tere  595 143088-143682         
#> 216        216         Dedibona Geferssa  775 143683-144457         
#> 217        217                    Ankabo  535 144458-144992         
#> 218        218               Mussie Ager  560 144993-145552         
#> 219        219              Lebona Derka  590 145553-146142       27
#> 220        220                  Ayedefer  620 146143-146762         
#> 221        221                  Abayelos  640 146763-147402         
#> 222        222                Tach kelad  630 147403-148032         
#> 223        223                  Tiququro  530 148033-148562         
#> 224        224                 Aba Alula  510 148563-149072         
#> 225        225                 Lay kelad  960 149073-150032         
#> 226        226                 wayu Ager  720 150033-150752         
#> 227        227                Dawud Ager  780 150753-151532       28
#> 228        228                  Enessana  840 151533-152372         
#> 229        229                    Jejero  605 152373-152977         
#> 230        230                     Kundi  580 152978-153557         
#> 231        231                  Zawuyoch  580 153558-154137         
#> 232        232                      Tilk  555 154138-154692         
#> 233        233                   Goshign  705 154693-155397         
#> 234        234         Geter(Enkulaloch)  550 155398-155947         
#> 235        235                 Teje Weha  630 155948-156577         
#> 236        236                 Aba balie  735 156578-157312       29
#> 237        237                    Ayteke  720 157313-158032         
#> 238        238              Eyasuna Jawi  615 150833-158647         
#> 239        239                    Bikeru  580 158648-159227         
#> 240        240                  Abajerie  580 159228-159807         
#> 241        241                  Aba bula  435 159808-160242         
#> 242        242                Aba Jareso  535 160243-160777         
#> 243        243               Amara Genda  630 160778-161407         
#> 244        244           Erefona cherkos  620 161408-162027         
#> 245        245      Legelafetona Alebeko  645 162028-162672       30
#> 246        246                   Amboasa  465 162673-163137         
#> 247        247   Kemerdengayna Menderoch  435 163138-163572         
#> 248        248                      Ambo  510 163573-164082         
#> 249        249                      Soni  525 164083-164607         
#> 250        250                  Birbirti  445 164608-165052         
#> 251        251                    Gedebo  690 165053-165742         
#> 252        252                      Meyo  520 165743-166262         
#> 253        253                   Sereser  480 166263-166742         
#> 254        254                      Rera  535 166743-167277         
#> 255        255                Legelafeto  530 167278-167807       31
#> 256        256           Chichitana kotu  400 167808-168207         
#> 257        257     AbarebaAgerna Shariya  570 168208-168777         
#> 258        258                     Beret  495 168778-169272         
#> 259        259                 Horu Ager  550 169273-169822         
#> 260        260                   Tatessa  620 169823-170442         
#> 261        261                  Aba Doyo  645 170443-171087         
#> 262        262                 Jeba Ager  555 171088-171642         
#> 263        263             Aba Jibo Ager  545 171643-172187         
#> 264        264                   Babogna  760 172188-172947         
#> 265        265                    Gadula  630 172948-173577       32
#> 266        266                   Kilkilo  555 173578-174132         
#> 267        267                  Gemeboba  490 174133-174622         
#> 268        268                    Satiro  515 174623-175137         
#> 269        269                 Adem Ager  535 175138-175672         
#> 270        270                     Dedie  445 175673-176117         
#> 271        271           Jaferena Teloch  445 176118-176562         
#> 272        272              Yimerie Ager  875 176563-177437         
#> 273        273                   Endulti  610 177438-178047         
#> 274        274                     Washa  870 178048-178917       33
#> 275        275                     Tucha  895 178918-179812         
#> 276        276                     Kundo  750 179813-180562         
#> 277        277                     Abayo  585 180563-181147         
#> 278        278                      Gero  765 181148-181912         
#> 279        279                    Wahilo  790 181913-182702         
#> 280        280                  Senselet  475 182702-183177         
#> 281        281                      Dego  945 183178-184122       34
#> 
#> $`Sample clusters 2`
#>    cluster_id    ctype village_id              village_name  pop       cum_pop
#> 1           1  primary         12                     Ayata 1070     4306-5375
#> 2           2  primary         17                 Tinibilka  550   10596-11145
#> 3           3  primary         25      Berara tuwate tsehay  915   15446-16360
#> 4           4  primary         31        Fela Tuwate tsehay  960   21441-22400
#> 5           5 reserved         37                    Meyane 1120   26901-28020
#> 6           6  primary         43             Assinefechign  850   32116-32965
#> 7           7  primary         51                 Chewagisa  870   37356-38225
#> 8           8  primary         57                    Guassa 1270   41931-43200
#> 9           9  primary         67                     Derka 1020   47761-48780
#> 10         10  primary         76                    Gergas 1075   53530-54605
#> 11         11  primary         82                     Arero  610   58951-59560
#> 12         12 reserved         89                 Werebecho  786   64460-65245
#> 13         13  primary         95                   Betassa  751   69990-70740
#> 14         14  primary        102                      Jema  751   75496-76246
#> 15         15  primary        110               Dabosh Ager  515   80764-81278
#> 16         16  primary        126 Dolie Ager and Abaya ager 1030   86228-87258
#> 17         17  primary        133               Abageferssa  480   91524-92003
#> 18         18  primary        141                    Tigaja  805   97189-97993
#> 19         19  primary        154                    Jejero  514 102536-103050
#> 20         20  primary        162                     Yegof  885 107568-108452
#> 21         21  primary        168                   Bekimos  750 113113-113862
#> 22         22  primary        177                    Abecho  770 118623-119392
#> 23         23 reserved        185                      Tiro  580 124203-124782
#> 24         24  primary        191          Doyenona Bukassa  720 129133-129852
#> 25         25  primary        200                  Abadeneb  450 134927-135377
#> 26         26  primary        210        Abiagerna Betoager  625 140488-141112
#> 27         27  primary        219              Lebona Derka  590 145553-146142
#> 28         28  primary        227                Dawud Ager  780 150753-151532
#> 29         29  primary        236                 Aba balie  735 156578-157312
#> 30         30  primary        245      Legelafetona Alebeko  645 162028-162672
#> 31         31 reserved        255                Legelafeto  530 167278-167807
#> 32         32  primary        265                    Gadula  630 172948-173577
#> 33         33  primary        274                     Washa  870 178048-178917
#> 34         34  primary        281                      Dego  945 183178-184122
#> 
#> $`Reserved clusters`
#> [1] 23 12 31  5
#>