Create a mouse object

mouse() constructs an S3 object of class 'mouse'. A mouse object consists of a list of slice objects and attributes stored as a list.

The slice objects are added to a mouse object with the function add_slice(). Each slice is a named element in the mouse object list, with the naming convention dependent on the slice ID and hemisphere attributes of the slice object.

If you are processing either a left or right hemisphere, the slice is named with the convention: "slice_ID" appended with its "hemisphere" If the hemisphere attribute is NULL, i.e. if the whole slice aligns well with a single atlas plate and there is no need to create separate slice objects per hemisphere, then the slice is named with the convention: "slice_ID"

Usage

mouse(
  mouse_ID = "set ID",
  sex = "female",
  age = NULL,
  genotype = NULL,
  reporter = NULL,
  strain = NULL,
  experiment = NULL,
  group = NULL,
  drug = NULL,
  cohort = NULL,
  input_path = "set input path",
  output_path = "set output path",
  ...
)

Arguments

mouse_ID

(str, default = 'set ID') e.g. '1_1'

sex

(str, default = "female")

age

(str, default = NULL)

genotype

(str, default = NULL)

reporter

(str, default = NULL)

strain

(str, default = NULL) e.g. 'B6'

experiment

(str, default = NULL) e.g. 'sundowning'

group

(str, default = NULL) e.g. 'control' or 'AD'

drug

(str, default = NULL) e.g. 'vehicle' or 'ketamine'

cohort

(str, default = NULL)

input_path

(str, default = 'set input path') Root to path containing the mouse data and slice image subfolders. This is useful if you've have changed computers and the drive mapped to the data has slightly changed. TODO: currently not used. Will include search function to parse out files and individual slice information for a mouse.

output_path

(str, default = 'set output path') Set the path to the folder you want to save the mouse RDATA file to.

...

additional custom keyword pair attributes you'd like to store

Value

A mouse, a colloquial term for an object of class 'mouse'. A 'mouse' object is also a list, with class list.

Details

The mouse attributes can be assigned as arguments to the mouse constructor function. See the parameters listed for the default values for these attributes Note that you are able to add custom attributes as keyword pairs, if you would like to keep track of an additional piece of information. However, this will only serve a descriptive purpose and will not be used for analysis. You may not need to use all mouse attributes but fill out as many are applicable to your experiment.

See also

Examples

mouse_example <- mouse() # initializes a mouse object